Adding makefile and adding the base game logic
Added makefile, which can compile for both macOS and Windows with commands make macOS and make windows, as well as adding logic for camera and moving with wasd, mouse rotates player towards the mouse, and the camera follows the player, the map is 1000 units. The player moves at 1.56 units per second using math and delta time.